move Loading to Common
