+++ client/src/components/Common/Loading/index.js
